requirements.txt
garminconnect>=0.2.0
garth>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
jinja2>=3.1.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
